# open_api/openapi.yaml
openapi: 3.0.3
info:
  title: FormulOne
  version: 1.0.0
servers:
- url: http://localhost:8080
  description: testing server
...
paths:
...
  /api/teams:
    get:
      tags:
        - teams
      summary: Get all teams
      description: Returns all teams
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Team'
        '500':
          description: Status internal server error

    post:
      tags:
        - teams
      summary: Create team
      description: ''
      requestBody:
        description: Create
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Team'
        required: true
      responses:
        '200':
          description: Successful operation
        '400':
          description: Invalid input
        '401':
          description: Unauthorized user
        '403':
          description: Not enough rights
        '500':
          description: Status internal server error
  /api/teams/{id}:
    get:
      tags:
        - teams
      summary: Get team
      description: Return team by id
      parameters:
        - name: id
          in: path
          description: ID of team
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Team'
        '400':
          description: Invalid input
        '404':
          description: Id doesnt exist
        '500':
          description: Status internal server error
    put:
      tags:
        - teams
      summary: Update team
      description: ''
      parameters:
        - name: id
          in: path
          description: ID of team
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Update team
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Team'
        required: true
      responses:
        '200':
          description: Successful operation
        '400':
          description: Invalid input
        '401':
          description: Unauthorized user
        '403':
          description: Not enough rights
        '404':
          description: Id doesnt exist
        '500':
          description: Status internal server error
    delete:
      tags:
        - teams
      summary: Delete team
      description: ''
      parameters:
        - name: id
          in: path
          description: ID of team
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
        '401':
          description: Unauthorized user
        '403':
          description: Not enough rights
        '404':
          description: Id doesnt exist
        '500':
          description: Status internal server error
  /api/tracks:
    get:
      tags:
        - tracks
      summary: Get all tracks
      description: Returns all tracks
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Track'
        '500':
          description: Status internal server error

    post:
      tags:
        - tracks
      summary: Create track
      description: ''
      requestBody:
        description: Create
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Track'
        required: true
      responses:
        '200':
          description: Successful operation
        '400':
          description: Invalid input
        '401':
          description: Unauthorized user
        '403':
          description: Not enough rights
        '500':
          description: Status internal server error
  /api/tracks/{id}:
    get:
      tags:
        - tracks
      summary: Get track
      description: Return track by id
      parameters:
        - name: id
          in: path
          description: ID of track
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Track'
        '400':
          description: Invalid input
        '404':
          description: Id doesnt exist
        '500':
          description: Status internal server error
    put:
      tags:
        - tracks
      summary: Update track
      description: ''
      parameters:
        - name: id
          in: path
          description: ID of track
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Update track
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Track'
        required: true
      responses:
        '200':
          description: Successful operation
        '400':
          description: Invalid input
        '401':
          description: Unauthorized user
        '403':
          description: Not enough rights
        '404':
          description: Id doesnt exist
        '500':
          description: Status internal server error
    delete:
      tags:
        - tracks
      summary: Delete track
      description: ''
      parameters:
        - name: id
          in: path
          description: ID of track
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
        '401':
          description: Unauthorized user
        '403':
          description: Not enough rights
        '404':
          description: Id doesnt exist
        '500':
          description: Status internal server error
...
components:
  schemas:
...
    Team:
      type: object
      properties:
        team_id:
          type: integer
          format: int64
          example: 121
        team_name:
          type: string
          example: RedBull
        team_country:
          type: string
          example: Hungary
        team_base:
          type: string
          example: Town
    Track:
      type: object
      properties:
        track_id:
          type: integer
          format: int64
          example: 121
        track_name:
          type: string
          example: RedBull
        track_country:
          type: string
          example: Hungary
        track_town:
          type: string
          example: Town
...